+++ src/block_average.h
@@ -7,17 +7,17 @@ typedef struct block_average_results {
   double *mean;
   double *var;
   double *blockvar;
-  int *nstride;
-  int *nvals;
-  unsigned int nblocks;
+  size_t *nstride;
+  size_t *nvals;
+  size_t nblocks;
 } block_average_results;
 
-void print_block_average_results (block_average_results *results, unsigned int vflag);
+void print_block_average_results (block_average_results *results, size_t vflag);
 
-block_average_results * block_average_results_alloc (unsigned int nblocks);
+block_average_results * block_average_results_alloc (size_t nblocks);
 
 void block_average_results_free (block_average_results *results);
 
-unsigned int block_average (unsigned int N, double *data, unsigned int nblocks, block_average_results *results);
+size_t block_average (size_t N, double *data, size_t nblocks, block_average_results *results);
 
 #endif
