+++ src/powers.h
@@ -1,11 +1,12 @@
 #ifndef __MYLIB_POWERS_H__
 #define __MYLIB_POWERS_H__
 
+#include <stdio.h>
 #include <math.h>
 #include <gsl/gsl_sf_log.h>
 
-unsigned int is_integer (double x);
+size_t is_integer (double x);
 
-unsigned int is_power_of_n (int x, int n);
+size_t is_power_of_n (int x, int n);
 
 #endif
